feat: integrate scheduled snapshots tests
