+++ packages/ats/contracts/contracts/facets/layer_1/snapshot/ISnapshots.sol
@@ -3,7 +3,7 @@ pragma solidity >=0.8.0 <0.9.0;
 
 import { ScheduledTask } from "../../layer_2/scheduledTask/scheduledTasksCommon/IScheduledTasksCommon.sol";
 
-// Snapshotted values have arrays of ids and the value corresponding to that id. These could be an array of a
+// Snapshot values have arrays of ids and the value corresponding to that id. These could be an array of a
 // Snapshot struct, but that would impede usage of functions that work on an array.
 struct Snapshots {
     uint256[] ids;
